Add `Java.g4` from `antlr/grammars-v4`.
This commit adds a hard-copy of the Antlr4 grammar written by Terence
Parr and Sam Harwell and made available on GitHub under the
`antlr/grammars-v4` repo under the BSD license.

The exact link is

    https://github.com/antlr/grammars-v4/blob/master/java/Java.g4

and it was made from commit

    7af37b609091557cbbf1e838d8dcdf833e1e6982

This grammar file was copied to our repo instead of somehow linking to
it (or the sources generated from it) because I did not immediately find
a clean and conventional way to do so.

The `pom.xml` was also updated to include the `antlr4-maven-plugin`.
This plugin is responsible for invoking Antlr4 to compiling a parser
from this grammar.